pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Никита Швейцер <shveitcer2023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standart"], version = "^0.24.0.post1"}
fastapi = {extras = ["all"], version = "^0.104.1"}
asyncpg = "^0.29.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
alembic = "^1.13.0"
aiosqlite = "^0.19.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.1"
python-multipart = "^0.0.6"
aiosmtplib = "^3.0.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}




[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
pytest-asyncio = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
    ".", "app",
]
asyncio_mode="auto"